Reorganized file structure for better project growth. Added alias import plugin for quality of life imporvements to importing. Added Layout component to house global styles and seo
