+++ package.json
@@ -17,6 +17,7 @@
     "ajv": "^6.10.0",
     "chart.js": "^2.8.0",
     "gatsby": "^2.19.10",
+    "gatsby-plugin-alias-imports": "^1.0.5",
     "gatsby-plugin-styled-components": "^3.0.7",
     "react": "^16.12.0",
     "react-chartjs-2": "^2.7.6",
