+++ src/modules/charts/ChartDoughnut.js
@@ -1,5 +1,5 @@
 import React from "react"
-import {doughnutData} from "../utils.js"
+import {doughnutData} from "@src/utils.js"
 import {Doughnut} from 'react-chartjs-2'
 import Chart from "chart.js";
