+++ src/modules/charts/PrimaryStatProjectCountries.js
@@ -1,15 +1,15 @@
 import React from "react"
 import styled from "styled-components"
-import { RowOverflow } from "../components/RowOverflow.js"
-import { region } from "../../static/data/dataformat.js"
-import { data } from "../utils.js"
-import { device } from "../mediaqueries/device.js"
-import imgAfrica from "../img/africa.svg"
-import imgAsia from "../img/asia.svg"
-import imgEurope from "../img/europe.svg"
-import imgNorthAmerica from "../img/northamerica.svg"
-import imgOceania from "../img/oceania.svg"
-import imgSouthAmerica from "../img/southamerica.svg"
+import { RowOverflow } from "@modules/ui/RowOverflow.js"
+import { region } from "@static/data/dataformat.js"
+import { data } from "@src/utils.js"
+import { device } from "@src/mediaqueries/device.js"
+import imgAfrica from "@images/africa.svg"
+import imgAsia from "@images/asia.svg"
+import imgEurope from "@images/europe.svg"
+import imgNorthAmerica from "@images/northamerica.svg"
+import imgOceania from "@images/oceania.svg"
+import imgSouthAmerica from "@images/southamerica.svg"
 
 // Geographic Representation
 const continentImages = {
