+++ src/modules/ui/Footer.js
@@ -1,8 +1,8 @@
 import React from "react"
 import styled from "styled-components"
-import { device } from "../mediaqueries/device.js"
-import { SectionWrapper } from "../components/SectionWrapper.js"
-import Label from "../components/Label.js"
+import { device } from "@src/mediaqueries/device.js"
+import { SectionWrapper } from "./SectionWrapper"
+import Label from "./Label"
 
 const SectionWrapperFooter = styled(SectionWrapper)`
   grid-template-columns: 1fr 1fr;
