+++ src/modules/ui/ProjectList.js
@@ -1,13 +1,13 @@
 import React from "react"
 import { Component } from "react"
 import styled from "styled-components"
-import { SectionWrapper } from "../components/SectionWrapper.js"
-import SectionTitle from "../components/SectionTitle.js"
-import { PrimaryButton } from "../components/PrimaryButton.js"
-import ProjectListRows from "../components/ProjectListRows.js"
-import { device } from "../mediaqueries/device.js"
-import { data, pieData } from "../utils.js"
-import { pType } from "../../static/data/dataformat.js"
+import { SectionWrapper } from "@modules/ui/SectionWrapper.js"
+import SectionTitle from "@modules/ui/SectionTitle.js"
+import { PrimaryButton } from "@modules/ui/PrimaryButton.js"
+import ProjectListRows from "@modules/ui/ProjectListRows.js"
+import { device } from "@src/mediaqueries/device.js"
+import { data, pieData } from "@src/utils.js"
+import { pType } from "@static/data/dataformat.js"
 import { Pie } from "react-chartjs-2"
 import Chart from "chart.js"
