+++ src/modules/ui/ProjectListRows.js
@@ -1,7 +1,7 @@
 import React from "react"
 import styled from "styled-components"
 import projectListRowStyles from "./projectlistrows.module.css"
-import { device } from "../mediaqueries/device.js"
+import { device } from "@src/mediaqueries/device.js"
 import { FaGithub } from "react-icons/fa"
 import { FaMedium } from "react-icons/fa"
 import { FaTwitter } from "react-icons/fa"
